+++ NEWS.md
@@ -1,3 +1,20 @@
+
+sate v2.1.0 (Release date: 2023-12-20)
+==============
+
+Major Changes:
+
+* Major revision of all existing functions. 
+* Graphing functions introduced, allowing users to plot the estimated probability
+  of guilty verdict in single trial condition, or plot the effect of a trial error
+  as difference between actual and hypothetical trial conditions.
+* Deliberation process modeled using a deliberate function which reduces standard
+  errors of estimates. Users have option to estimate values using empirical model
+  and standard errors generated through computational methods.
+* Account for selection of small trial jury from large jury pool.
+* Introduced optional arguments to account for varying jury size and peremptory strikes.
+
+
 sate v1.1.0 (Release date: 2020-03-17)
 ==============